Millersville holdings locations transformations (#21)
* Clean up XSL for library codes mapping

* Prevent duplicated holdings and items, use right location field

  The 900 tag (holdings record) sometimes appears twice in the
  record with the same holdings record ID. This caused the holdings
  to be duplicated together with all its items.

  Now one of them is ignored - by its indicator value ('4') - but this
  approach is to be verified.

  In case it results in no holdings record being forwarded, this
  condition must be caught in InventoryRecordStorage (implemented too)

  Also uses subfield 'd' rather than 'b' for the location code

+++ harvester/src/main/java/com/indexdata/masterkey/localindices/harvest/storage/InventoryRecordStorage.java
@@ -398,31 +398,34 @@ public class InventoryRecordStorage implements RecordStorage {
       newHoldingsRecord = (JSONObject) holdingsRecords.get(0);
     } catch (ClassCastException cce) {
       logger.warn("Could not get holdingsRecord object from " + holdingsRecords.get(0) + " " + cce.getMessage());
+    } catch (IndexOutOfBoundsException ioobe) {
+      logger.error("Found incoming holdingsRecords element but no holdings records in it. Will skip deleting any existing holdings records");
     }
-
-    JSONArray existingHoldingsRecords = getHoldingsRecordsByInstanceId(instanceId);
-    if (existingHoldingsRecords != null) {
-      Iterator<JSONObject> existingHoldingsRecordsIterator = existingHoldingsRecords.iterator();
-      while (existingHoldingsRecordsIterator.hasNext()) {
-        JSONObject existingHoldingsRecord = existingHoldingsRecordsIterator.next();
-        String existingHoldingsRecordId = (String) existingHoldingsRecord.get("id");
-        if (fromSameInstitution(existingHoldingsRecord, newHoldingsRecord)) {
-          JSONArray items = getItemsByHoldingsRecordId(existingHoldingsRecordId);
-          if (items != null) {
-            Iterator<JSONObject> itemsIterator = items.iterator();
-            while (itemsIterator.hasNext()) {
-              JSONObject item = itemsIterator.next();
-              String itemId = (String) item.get("id");
-              deleteItem(itemId);
+    if (newHoldingsRecord != null) {
+      JSONArray existingHoldingsRecords = getHoldingsRecordsByInstanceId(instanceId);
+      if (existingHoldingsRecords != null) {
+        Iterator<JSONObject> existingHoldingsRecordsIterator = existingHoldingsRecords.iterator();
+        while (existingHoldingsRecordsIterator.hasNext()) {
+          JSONObject existingHoldingsRecord = existingHoldingsRecordsIterator.next();
+          String existingHoldingsRecordId = (String) existingHoldingsRecord.get("id");
+          if (fromSameInstitution(existingHoldingsRecord, newHoldingsRecord)) {
+            JSONArray items = getItemsByHoldingsRecordId(existingHoldingsRecordId);
+            if (items != null) {
+              Iterator<JSONObject> itemsIterator = items.iterator();
+              while (itemsIterator.hasNext()) {
+                JSONObject item = itemsIterator.next();
+                String itemId = (String) item.get("id");
+                deleteItem(itemId);
+              }
             }
+            deleteHoldingsRecord(existingHoldingsRecordId);
+          } else {
+            logger.debug("holdingsRecord " + existingHoldingsRecordId + " belongs to a different institution (" + existingHoldingsRecord.get("permanentLocationId") +"), not deleting it.");
           }
-          deleteHoldingsRecord(existingHoldingsRecordId);
-        } else {
-          logger.debug("holdingsRecord " + existingHoldingsRecordId + " belongs to a different institution (" + existingHoldingsRecord.get("permanentLocationId") +"), not deleting it.");
         }
+      } else {
+        logger.info("No existing holdingsRecords found for the instance, nothing to delete.");
       }
-    } else {
-      logger.info("No existing holdingsRecords found for the instance, nothing to delete.");
     }
   }
